pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Chirag-droid <chirag.singla.pi@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.6"
"discord.py" = "^1.7.3"
PyYAML = "^5.4.1"
arrow = "^1.1.1"
"nekos.py" = "^1.0.3"
requests = "^2.26.0"
pymongo = {extras = ["srv"], version = "^3.12.0"}
motor = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
flake8 = "^3.9.2"
black = "^21.6b0"
pre-commit = "^2.13.0"
taskipy = "^1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = { cmd = "python -m discord_bot", help = "Runs the bot"}
lint = { cmd = "pre-commit run --all-files", help = "Lints code using black and flake8" }
precommit = { cmd = "pre-commit install", help = "Sets up pre-commit for the project" }
